Plan-Do-Agent/pyproject.toml
[project]
name = "plan-do-agent"
version = "0.1.0"
description = "A Plan-and-Execute agent with persistent todo list using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/plan_do_agent"]
